+++ RunningStatistics/IRunningStatistic.cs
@@ -5,7 +5,7 @@ namespace RunningStatistics;
 
 public interface IRunningStatistic<TObs>
 {
-    public nint Count { get; }
+    public long Count { get; }
 
     public void Fit(IEnumerable<TObs> values);
